updated abstraction and descriptions
+ updated abstraction of sql
+ added descriptions to new classes
