+++ src/main/java/xyz/geik/farmer/database/DatabaseType.java
@@ -1,5 +1,10 @@
 package xyz.geik.farmer.database;
 
+/**
+ * DatabaseType enum class
+ * @since b000
+ * @author Amowny
+ */
 public enum DatabaseType {
     SQLITE,
     MYSQL;
