+++ src/main/resources/storage/database.yml
@@ -1,9 +1,19 @@
+# Database configurations
+# If you don't want to setup this plugin to MySQL then leave it default
 database:
+  # type of sql (mysql | sqlite) default: sqlite
   type: sqlite # mysql or sqlite
+  # ip of mysql default: 127.0.0.1
   host: 127.0.0.1
+  # port of mysql default: 3306
   port: 3306
+  # username of sql default: root
   username: root
+  # password of sql default: <no any default>
   password: "ilovegeikplugins"
+  # database name of mysql default: <no any default>
   database: geik_farmer
+  # table of farmer in sql default: farmers
   farmers_table: farmers
+  # users table of farmer in sql default: farmer_users
   farmer_users_table: farmer_users
